Setting vocabulary datasets default batch_size to tf.int32.max instead of tf.int64.max, because the latter would cause a runtime error converting the value to `long long`.
PiperOrigin-RevId: 322177556

+++ tensorflow_transform/tf_utils.py
@@ -463,7 +463,7 @@ class _DatasetInitializerCompat(
     return init_op
 
 
-_DATASET_BATCH_SIZE = tf.int64.max
+_DATASET_BATCH_SIZE = tf.int32.max
 
 
 def read_tfrecord_vocabulary_dataset(vocab_path):
